Get latest version of chai-kefir
2.0.0 now exports a factory function which returns the helpers
we need.
