+++ packages/brookjs/package.json
@@ -49,7 +49,7 @@
     "brookjs-desalinate": "0.0.0",
     "chai": "^4.1.2",
     "chai-dom": "^1.4.0",
-    "chai-kefir": "^1.1.0",
+    "chai-kefir": "^2.0.1",
     "chai-spies": "^0.7.1",
     "core-js": "^2.4.1",
     "cross-env": "^5.0.1",
